foundations_rest_api/src/foundations_rest_api/v1/models/queued_job.py
"""
Copyright (C) DeepLearning Financial Technologies Inc. - All Rights Reserved
Unauthorized copying, distribution, reproduction, publication, use of this file, via any medium is strictly prohibited
Proprietary and confidential
Written by Thomas Rogers <t.rogers@dessa.com>, 06 2018
"""

from foundations_rest_api.v1.models.property_model import PropertyModel


class QueuedJob(PropertyModel):

    job_id = PropertyModel.define_property()
    user = PropertyModel.define_property()
    submitted_time = PropertyModel.define_property()

    @staticmethod
    def all():
        from foundations_rest_api.response import Response

        def _all():
            return []

        return Response('QueueJob', _all)
        